This fixes the Viewport re-sizing issue for ExtendViewport.
Lines 59 through 64 fixes the Viewport re-sizing issue from 1.9.10 and before, concerning ExtendViewport and OrthographicCamera.  In certain situations where an ExtendViewport object was quickly removed and created anew, it would double the values of the OrthographicCamera reacting to the parent Viewport class variables of worldWidth and worldHeight.  The OrthographicCamera would effectively act as if zoomed in by 200%.

Lines 59 through 64 verify if the Camera type is truly an OrthographicCamera and reacts accordingly with the fix.  This may carry a benefit for future 3D endeavors with non-Orthographic Camera types, such as PerspectiveCamera.

I have tested this with my own code, relying upon my own version of ExtendViewport only.  Thank you for reading.